refactor: split files over the 500-line ratchet
Keep experimental tool prefs, source control root, and related tests in dedicated files so static analysis stays under the max-lines check.
