+++ lib/src/features/workbench/application/workbench_controller_internals.dart
@@ -453,29 +453,7 @@ mixin _WorkbenchControllerInternals on _$WorkbenchController {
 
   void _focusExperimentalTerminal(String workspaceId, String? key);
 
-  void _seedExperimentalNewWorkspacePanel(String workspaceId) {
-    if (!state.isExperimentalLayout) {
-      return;
-    }
-    final tools = ExperimentalWorkspaceTool.uniqueInOrder(
-      state.viewPrefs.experimentalNewWorkspaceTools,
-    );
-    if (tools.isEmpty) {
-      return;
-    }
-    final panel = state.experimentalPanelFor(workspaceId);
-    final alreadyHasTools = panel.occupiedKeys.any(
-      (key) => ExperimentalWorkspaceTool.forKey(key) != null,
-    );
-    if (alreadyHasTools) {
-      return;
-    }
-    _saveExperimentalPanel(
-      workspaceId,
-      panel.openToolsInOrder(tools),
-      reveal: true,
-    );
-  }
+  void _seedExperimentalNewWorkspacePanel(String workspaceId);
 
   void _saveExperimentalPanel(
     String workspaceId,
